made webserver runnable from cli, added docs: laminar
