+++ pkg/codex/docs/coding-guide.md
@@ -6,6 +6,7 @@
 * Dream > Document > Design > Demo > Deliver 
 
 **Table of Contents** 
+
 - [Configurations](#configuration)
 - [Guidelines](#guidelines)
    - [Testing](#testing)
