+++ pkg/codex/docs/git.md
@@ -1,0 +1,34 @@
+# Git 
+
+## Commands
+
+### Setup
+- `git config --global user.name "[firstname lastname]"`
+- `git config --global user.email "[valid-email]"`
+- `git init`
+- `git clone`
+
+### Stage and Snapshot
+- `git status`
+- `git add [file]`
+- `git reset [file]`
+- `git diff`
+- `git diff --staged`
+- `git commit -m "[message]"`
+
+### Branching
+
+### Log
+- `git log`
+- `git log branchA..branchB`
+- `git log --follow [file]` 
+
+### Rewrite History
+- `git rebase [branch]`
+- `git reset --hard [commit]`
+
+### Stash
+- `git stash` `list | pop | drop`
+
+## References 
+- https://education.github.com/git-cheat-sheet-education.pdf
