+++ pkg/codex/docs/javascript.md
@@ -1,0 +1,28 @@
+<h1 class="counter-display">(..)</h1>
+<button class="counter-minus">-</button>
+<button class="counter-plus">+</button>
+
+<script>
+
+let counterDisplayElem = document.querySelector('.counter-display');
+let counterMinusElem = document.querySelector('.counter-minus');
+let counterPlusElem = document.querySelector('.counter-plus');
+
+let count = 0;
+
+updateDisplay();
+
+counterPlusElem.addEventListener("click",()=>{
+    count++;
+    updateDisplay();
+}) ;
+
+counterMinusElem.addEventListener("click",()=>{
+    count--;
+    updateDisplay();
+});
+
+function updateDisplay(){
+    counterDisplayElem.innerHTML = count;
+};
+</script>
