+++ pkg/codex/docs/laminar.md
@@ -1,0 +1,20 @@
+# Laminar for ScalaJS
+
+## Usage
+
+build.sbt
+```
+"com.raquo" %%% "laminar" % "0.13.1"  // Requires Scala.js >= 1.5.0
+"com.raquo" %%% "airstream" % "<version>"
+```
+
+import
+```
+import com.raquo.api.L.{*, given}  // Scala 3
+```
+
+
+## References
+- [https://github.com/raquo/Laminar/](https://github.com/raquo/Laminar/) - Under [MIT License](https://github.com/raquo/laminar/blob/master/LICENSE.md).
+- [Laminar Official Docs](https://laminar.dev/documentation)
+- [Laminar - Smooth UI Youtube](https://www.youtube.com/watch?v=L_AHCkl6L-Q)
